Added files for the gimik program.
added main.py which is responsible for running the main program.
added user_interact.py which will contain methods for user to
use to interact with the cad program.
added view_gimik.py which will be responsible for creating the
3D view of the shapes.
added workspace.py which is reponsible for storing all the shape
information.